+++ packages/antd-ext/components/ButtonExt/Mix.tsx
@@ -1,51 +1,32 @@
-import React, { useMemo, useContext } from "react";
-import { theme } from "antd";
+import React from "react";
 import ButtonExt from "./ButtonExt";
 import type { ButtonExtProps } from "./ButtonExt";
 import {
   Theme as ThemeHermes,
-  customStyle as customStyleHermes,
-  ThemeProps
-} from "./ThemeHermes";
-import classNames from "classnames";
-import { AntdExtGlobalContext } from '../ConfigProviderExt/context';
+  customStyle as customStyleHermes} from "./ThemeHermes";
 import type { Theme } from '../ConfigProviderExt/context';
-import { ConfigProvider } from 'antd';
-
-const { useToken } = theme;
+import useMapTheme from "../utils/useMapTheme";
 
 export interface ButtonExtMinixProps extends ButtonExtProps {
   theme?: Theme;
 }
 
 export default function Minix(props: ButtonExtMinixProps) {
-  const { token } = useToken();
-  const { themeExt } = useContext(AntdExtGlobalContext);
-  const { getPrefixCls } = React.useContext(ConfigProvider.ConfigContext);
-  const { className, theme: customTheme, ...restProps } = props;
-
-  const classes = useMemo(
-    () => (
-      {
-        hermes: classNames(className, [customStyleHermes(token, getPrefixCls())]),
-        zeus: classNames(className),
-        default: classNames(className),
-      }[customTheme || themeExt]
-    ),
-    [className, token, customTheme, themeExt]
-  );
-
-  const Theme = useMemo(() => (
-    {
+  const { className, theme, ...restProps } = props;
+  const { classes, ThemeWrapper } = useMapTheme({
+    className,
+    theme,
+    themeWrap: {
       hermes: ThemeHermes,
-      zeus: (props: ThemeProps) => <>{props.children}</>,
-      default: (props: ThemeProps) => <>{props.children}</>,
-    }[customTheme || themeExt]
-  ), [customTheme, themeExt]);
+    },
+    emotioncss: {
+      hermes: customStyleHermes,
+    }
+  });
 
   return (
-    <Theme>
+    <ThemeWrapper>
       <ButtonExt className={classes} {...restProps} />
-    </Theme>
+    </ThemeWrapper>
   );
 };