+++ packages/antd-ext/components/InputExt/index.ts
@@ -1,0 +1,4 @@
+import { MixinInputExt, MixinInputOutLineExt } from './Mix';
+import InputOutLineExt from './InputOutLineExt';
+
+export { InputOutLineExt, MixinInputExt, MixinInputOutLineExt };