+++ packages/antd-ext/components/index.ts
@@ -1,6 +1,7 @@
 import ButtonExt from './ButtonExt';
 import ConfigProviderExt from './ConfigProviderExt';
 import TableExt from './TableExt';
+import { MixinInputExt as InputExt, MixinInputOutLineExt as InputOutLineExt } from './InputExt';
 export * from 'antd';
 import zh_CN from 'antd/locale/zh_CN';
 
@@ -8,6 +9,8 @@ export {
   ButtonExt,
   ConfigProviderExt,
   TableExt,
+  InputExt,
+  InputOutLineExt,
 }
 
 export const locale = {